Improved the score upticks to be more dynamic

+++ core/src/main/java/github/team42/ggj25/entity/ScoreBoard.java
@@ -10,7 +10,7 @@ import github.team42.ggj25.Constants;
 import github.team42.ggj25.Drawable;
 
 public class ScoreBoard implements Drawable {
-    private int score = 0;
+    private long score = 0;
 
     private final Label scoreBoardLabel;
     private final Label.LabelStyle scoreBoardStyle = new Label.LabelStyle();
@@ -43,17 +43,17 @@ public class ScoreBoard implements Drawable {
         scoreBoardLabel.setText(score + " QUAKS");
     }
 
-    public int addPointsToScore(int points) {
+    public long addPointsToScore(long points) {
         score += points;
         return score;
     }
 
-    public int subtractPointsFromScore(int points) {
+    public long subtractPointsFromScore(long points) {
         score -= points;
         return score;
     }
 
-    public int getScore() {
+    public long getScore() {
         return score;
     }
 }
